Update get_reaction_consistencies.R
As discussed last time, there is -log2(x) in the source code but not -log2(1+x), this may be necessary to change in case of 0 value.
Since each row represents a reaction while columns represent samples. From the source code, we can see in the scale step, each value minus the mean of the column may affect the difference between different sample groups and then affect the statistical test between different sample groups when we identify the differential reactions. Why not minus the minimum value of each row? Maybe this is your original intention.